[MOD] trans: 번역 모듈 함수 이동
